requirements.txt
slack-bolt>=1.18.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-ollama>=0.0.1
